rfe12658: send-letter: allow alternate transport
<release-note>
The first argument to net.post-office:send-letter, the server
specification, can now specify an alternate transport mechanism,
instead of the usual SMTP network protocol.
</release-note>

<documentation>
net.post-office:send-letter's server argument can now be used to
specify a program that does the delivery of the message.  As in:

  (:program . "/usr/sbin/sendmail -t")
or
  (:program . ("/usr/sbin/sendmail" "-odq" "-t"))

The first form will invoke "/usr/sbin/sendmail -t" and pipe the message,
constructed from the other arguments to send-letter, to the program.
The second form will invoke excl:run-shell-command with this first
argument:

  (vector "/usr/sbin/sendmail" "/usr/sbin/sendmail" "-odq" "-t")

which allows the program to be executed without an intermediate shell.
</documentation>

Change-Id: I7811b99bee6819dee7e026ba464974bd2ea803d8